unit tests [stage-20]
