+++ src/app/template-editor/template-components/services/component-utils.service.spec.ts
@@ -1,3 +1,4 @@
+import { expect } from 'chai';
 import { TestBed } from '@angular/core/testing';
 
 import { ComponentUtilsService } from './component-utils.service';
@@ -11,6 +12,20 @@ describe('ComponentUtilsService', () => {
   });
 
   it('should be created', () => {
-    expect(service).toBeTruthy();
+    expect(service).to.exist;
   });
+
+  describe('isValidUrl', () => {
+    it('reports valid urls', () => {
+      expect(service.isValidUrl('www.risevision.com')).to.be.true;
+      expect(service.isValidUrl('https://www.risevision.com')).to.be.true;
+      expect(service.isValidUrl('risevision.com')).to.be.true;
+    });
+
+    it('flags invalid urls', () => {
+      expect(service.isValidUrl('.risevision.com')).to.be.false;
+      expect(service.isValidUrl('https://')).to.be.false;
+      expect(service.isValidUrl('risevision')).to.be.false;
+    });
+  })
 });
